delete unpaid existing payment transaction implemented - unit test - OK

+++ backend/src/main/java/com/pymntprocessing/pymntprocessing/repository/PaymentTransactionRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PaymentTransactionRepository extends JpaRepository<PaymentTransaction, Integer> {
+public interface PaymentTransactionRepository extends JpaRepository<PaymentTransaction, Long> {
 
     Optional<PaymentTransaction> findById(Long id);
     List<PaymentTransaction> findAllByVendorId(Long id);
